Refactored the structure
