+++ lib/core/styles/colors.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+const black=Colors.black;
+const blue=Colors.blue;
+const purple=Colors.purple;
+const pink=Colors.pink;
+const white=Colors.white;
+const green=Colors.green;
+const red=Colors.red;
+const grey=Colors.grey;
